refactor: Lotto에 있던 checkRank를 WinningNumbers로 옮김
