+++ src/test/java/lotto/model/domain/LottoTest.java
@@ -8,9 +8,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
 class LottoTest {
 
@@ -49,31 +46,5 @@ class LottoTest {
 
         // when & then
         assertThrows(IllegalArgumentException.class, () -> new Lotto(numbers));
-    }
-
-    @ParameterizedTest
-    @MethodSource("로또와_당첨번호_및_보너스볼을_비교해_Rank_반환_성공_테스트케이스")
-    void 로또와_당첨번호_및_보너스볼을_비교해_Rank_반환_성공(Lotto lotto, Rank expectedRank) {
-        // given
-        WinningNumbers winningNumbers = new WinningNumbers(
-                new Lotto(List.of(1, 2, 3, 4, 5, 6)),
-                LottoNumber.of(7));
-
-        // when
-        Rank rank = lotto.checkRank(winningNumbers);
-
-        // then
-        assertThat(rank).isEqualTo(expectedRank);
-    }
-
-    static Stream<Arguments> 로또와_당첨번호_및_보너스볼을_비교해_Rank_반환_성공_테스트케이스() {
-        return Stream.of(
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 6)), Rank.FIRST),
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 7)), Rank.SECOND),
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 8)), Rank.THIRD),
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 8, 9)), Rank.FOURTH),
-                Arguments.of(new Lotto(List.of(1, 2, 3, 8, 9, 10)), Rank.FIFTH),
-                Arguments.of(new Lotto(List.of(8, 9, 10, 11, 12, 13)), Rank.NONE)
-        );
     }
 }
