+++ src/test/java/lotto/model/domain/WinningNumbersTest.java
@@ -1,10 +1,15 @@
 package lotto.model.domain;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.List;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 class WinningNumbersTest {
 
@@ -24,5 +29,32 @@ class WinningNumbersTest {
         // when, then
         assertThrows(IllegalArgumentException.class,
                 () -> new WinningNumbers(new Lotto(winningNumbers), LottoNumber.of(6)));
+    }
+
+
+    @ParameterizedTest
+    @MethodSource("로또와_당첨번호_및_보너스볼을_비교해_Rank_반환_성공_테스트케이스")
+    void 로또와_당첨번호_및_보너스볼을_비교해_Rank_반환_성공(Lotto lotto, Rank expectedRank) {
+        // given
+        WinningNumbers winningNumbers = new WinningNumbers(
+                new Lotto(List.of(1, 2, 3, 4, 5, 6)),
+                LottoNumber.of(7));
+
+        // when
+        Rank rank = winningNumbers.checkRank(lotto);
+
+        // then
+        assertThat(rank).isEqualTo(expectedRank);
+    }
+
+    static Stream<Arguments> 로또와_당첨번호_및_보너스볼을_비교해_Rank_반환_성공_테스트케이스() {
+        return Stream.of(
+                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 6)), Rank.FIRST),
+                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 7)), Rank.SECOND),
+                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 8)), Rank.THIRD),
+                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 8, 9)), Rank.FOURTH),
+                Arguments.of(new Lotto(List.of(1, 2, 3, 8, 9, 10)), Rank.FIFTH),
+                Arguments.of(new Lotto(List.of(8, 9, 10, 11, 12, 13)), Rank.NONE)
+        );
     }
 }
